Ping endpoint for Paul
Return JSON with basic application information.
