+++ app/controllers/ping_controller.rb
@@ -1,4 +1,7 @@
 class PingController  < ApplicationController
+  respond_to :json
+
   def index
+    respond_with Deployment.info
   end
 end
