+++ app/views/layouts/moj_template.html.erb
@@ -82,7 +82,7 @@
             <%= yield :inside_header %>
           </div>
           <% if content_for?(:proposition_header) %>
-            <%= yield(:proposition_header) %>         
+            <%= yield(:proposition_header) %>
           <% else %>
             <div class="header-proposition">
               <div class="content">
@@ -118,7 +118,7 @@
 
       <footer class="group js-footer" id="footer" role="contentinfo">
         <div class="footer-wrapper">
-          <div><%= version_number %></div>
+          <div><%= Deployment.version_number %></div>
           <%= yield :footer_top %>
           <div class="footer-meta">
             <div class="footer-meta-inner">
