--- app/views/ping/index.html.erb
@@ -1,7 +1,0 @@
-<div>
-<p><i class="fa fa-smile-o"></i> <%= DateTime.now %></p>
-<p><i> Version Number </i> <%= version_number %></p>
-<p><i> Build Time </i> <%= build_time %></p>
-<p><i> Commit Id </i> <%= commit_id %></p>
-</div>
-
